Change output sections memory region assignments
Due to memory constraints, .data and .bss output sections are now
assigned to the larger ahb_sram_64k memory region. Due to lack of
configuration in the OpenOCD script file for the edu-ciaa-nxp board
the flash_bank_b is not flashable, so both .text and .rodata output
sections are now assigned to flash_bank_a until script file gets
upgraded.